+++ modules/logs/variables.tf
@@ -46,7 +46,7 @@ variable "log_analytics_workspace_retention_in_days" {
   default     = 30
 }
 
-variable "logs_storage_account_creation" {
+variable "logs_storage_account_enabled" {
   description = "Whether the dedicated Storage Account for logs is created."
   type        = bool
   default     = true