+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 Added
-  * AZ-1093: Add `logs_storage_account_creation` option to `logs` sub-module
+  * AZ-1093: Add `logs_storage_account_enabled` option to `logs` submodule
 
 # v7.3.2 - 2023-06-02