+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# Unreleased
+# v7.4.0 - 2023-06-16
 
 Added
   * AZ-1093: Add `logs_storage_account_enabled` option to `logs` submodule
+  * AZ-673: Patch Management Monitoring with Fame
 
 # v7.3.2 - 2023-06-02
 
@@ -20,7 +21,6 @@ Added
   * AZ-1070: Add `vm_backup_policy_type` new parameter
   * AZ-1055: Add `client_name` in Storage Account name. Use `var.logs_storage_account_custom_name` to keep your storage when upgrading.
   * AZ-1056: Bump `function-app` module for monitoring function
-  * AZ-673: Patch Management Monitoring with Fame
 
 Breaking
   * AZ-1038: Bump Key Vault module. Use `var.keyvault_custom_name` to keep your Key Vault when upgrading